Beta

+++ src/main/java/ru/spbstu/neer2015/Data/ParametrSelectionSetter.java
@@ -4,13 +4,13 @@ package ru.spbstu.neer2015.data;
  * Created by tseyler on 18.05.15.
  */
 public class ParametrSelectionSetter {
-    private double startC = 1;
-    private double endC = 100;
-    private double stepC = 1;
+    private int startC = 1;
+    private int endC = 100;
+    private int stepC = 1;
     private double startKernel = 1;
     private double endKernel = 4;
     private double stepKernel = 0.1;
-    public ParametrSelectionSetter(double startC, double endC, double stepC, double startKernel, double endKernel, double stepKernel) {
+    public ParametrSelectionSetter( int startC, int  endC, int stepC, double startKernel, double endKernel, double stepKernel) {
         this.startC = startC;
         this.endC = endC;
         this.stepC = stepC;
